+++ R/aaa.R
@@ -1,0 +1,4 @@
+#' @useDynLib simplecall, .registration=TRUE
+NULL
+
+