+++ src/create-data-frame.c
@@ -1,118 +1,11 @@
----
-output: github_document
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.path = "man/figures/README-",
-  out.width = "100%"
-)
-
-library(simplecall)
-```
-
-# simplecall - demo package for calling C code with `.Call()`
-
-<!-- badges: start -->
-![](https://img.shields.io/badge/cool-useless-green.svg)
-<!-- badges: end -->
-
-`simplecall` is a small demo package showing how C code could be included
-in a package and called with `.Call()`.
-
-This package now includes demo code for:
-
-* creating a list in C
-* creating a data.frame in C
-
-
-You can install from [GitHub](https://github.com/coolbutuseless/simplecall) with:
-
-``` r
-# install.package('remotes')
-remotes::install_github('coolbutuseless/simplecall)
-```
-
-## What's in the box?
-
-Package contains example C functions and R wrappers to call them
-
-| C function                                | R function   |
-|-------------------------------------------|--------------|
-| `SEXP add_(SEXP x_, SEXP y_)`             | `add_Call(x, y)`  |
-| `SEXP create_list_in_c_()`                | `create_list_in_c()` |
-| `SEXP create_data.frame_in_c_()`          | `create_data_frame_in_c()` |
 
 
 
-## What's in the R functions?
-
-* A call using `.Call()`
-* First argument is the C function name e.g. `add_`
-* Other arguments to the R function can be passed through to C 
-
-```{r eval = FALSE}
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Add two numbers
-#'
-#' @param x,y numbers to add
-#'
-#' @useDynLib simplecall add_
-#' @export
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-add_Call <- function(x, y) {
-  .Call(add_, x, y)
-}
-```
+#include <R.h>
+#include <Rinternals.h>
+#include <Rdefines.h>
 
 
-## What's in the C functions?
-
-* C function returns an `SEXP` pointer which is a real R object
-* all arguments passed as SEXP pointers
-
-```
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Add 2 numbers
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP add_(SEXP x_, SEXP y_) {
-
-  int xtype = TYPEOF(x_);
-  int ytype = TYPEOF(y_);
-
-  if (xtype != REALSXP | ytype != REALSXP) {
-    Rf_error("add_() expects only floating point numbers i.e. doubles or reals");
-  }
-
-  int xlen = length(x_);
-  int ylen = length(y_);
-
-  if (xlen != 1 | ylen != 1) {
-    error("add_() expects arguments to be length = 1");
-  }
-
-  SEXP result = PROTECT(allocVector(REALSXP, 1));
-  REAL(result)[0] = asReal(x_) + asReal(y_);
-  UNPROTECT(1);
-
-  return result;
-}
-```
-
-
-
-## Creating a R List within C
-
-Example code is included to create a list completely within C and return it to R
-
-<details>
-<summary> Click to show/hide code </summary>
-
-```{c eval = FALSE}
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Creating a list within C and returning it to R
 //
@@ -167,19 +60,11 @@ SEXP create_list_in_c_() {
   UNPROTECT(5);
   return list_;
 }
-```
-
-</details>
 
 
-## Creating a R data.frame within C
 
-Example code is included to create a data.frame completely within C and return it to R
 
-<details>
-<summary> Click to show/hide code </summary>
 
-```{c eval = FALSE}
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Creating a data.frame within C and returning it to R
 //
@@ -252,37 +137,3 @@ SEXP create_data_frame_in_c_() {
   UNPROTECT(6);
   return df_;
 }
-```
-
-</details>
-
-
-
-
-### Similar packages
-
-This is one of a series of small demo packages for  
-calling other languages from R:
-
-* [{simplec}](https://github.com/coolbutuseless/simplec) - calling C with `.C()`
-* [{simplecall}](https://github.com/coolbutuseless/simplecall) - calling C with `.Call()`
-* [{simplercpp}](https://github.com/coolbutuseless/simplercpp) - calling C++ with `{Rcpp}`
-* [{simplefortran}](https://github.com/coolbutuseless/simplefortran) - calling Fortran with `.Fortran()`
-
-## Installation
-
-
-## Resources
-
-* [David Vaughan's post - Now you C me](https://blog.davisvaughan.com/2019/03/02/now-you-c-me/)
-* [Peter Dalgaard's Keynote from UseR 2004](http://www.ci.tuwien.ac.at/Conferences/useR-2004/Keynotes/Dalgaard.pdf) 
-discusses R's language interfaces
-* [Using R — Calling C Code ‘Hello World’](http://mazamascience.com/WorkingWithData/?p=1067)
-* [Hadley's Advanced R book chapter - 'Rs C interface'](http://adv-r.had.co.nz/C-interface.html)
-* [Rcpp](https://cran.r-project.org/package=Rcpp)
-
-## Acknowledgements
-
-* R Core for developing and maintaining such a wonderful language.
-* CRAN maintainers, for patiently shepherding packages onto CRAN and maintaining
-  the repository