+++ NEWS.md
@@ -1,3 +1,11 @@
+
+# simplecall 0.1.1
+
+* Automatic registration of C functions.  See `R.aaa.R`
+* Added demonstration functions for:
+    * Generating a list within C
+    * Generating a data.frame within C
+
 # simplecall 0.1.0
 
 * Initial release